api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery==5.4.0
redis==5.0.7
requests==2.32.3
python-dotenv==1.0.1
openai-whisper==20231117
aiofiles==24.1.0
moviepy==1.0.3
yt-dlp==2024.10.22
structlog==24.1.0
pydantic==2.8.2
pydantic-settings==2.4.0
gTTS==2.5.1
deep-translator==1.11.4
pytest==8.3.2
ruff==0.6.8
black==24.8.0
